api/pyproject.toml
[tool.pytest.ini_options]
testpaths = ["./tests"]
filterwarnings = ["ignore::DeprecationWarning:.*datetime.datetime.*"]

[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Lillith Long lilli@adeptexec.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
pydantic = "^2.10.2"
odmantic = "^1.0.2"
pytest = "^8.3.3"
motor = "^3.6.0"
black = "^24.10.0"
flake8 = "^7.1.1"
typing-extensions = "^4.12.2"
uvicorn = "^0.32.1"
bcrypt = "^4.1.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
httpx = "^0.27.0"
PyJWT = "^2.8.0"
google-auth-oauthlib = "^1.2.0"
google-auth = "^2.28.2"
google-api-python-client = "^2.123.0"
cryptography = "^42.0.5"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"